CircleCI 2.0 (#91)
* Testing out CircleCI 2.0

* Add workflow

* Test template itself

* Better tests

* Set chruby

* cache cocoapods

* fix caching

* add python deps

* cache cocoapods

* Fix cookiecutter

* Fix cookiecutter

* Fix bundler

* Fix caching

* Fix caching again

* Fix build issues

* Add CircleCI badges

* Update Gemfile.lock

* Download CocoaPods from CircleCI directly

* Use Xcode 9.3

* Ruby 2.5

* cleanup circleci

* Revert to Xcode 9.2.0

* Ruby 2.4 again

* Pin Swiftilities

* Cache/restore Pods cache

* Fix generated template

* Use local template

* Use different gem caching

* Unpin Siwfiltities again

* remove unneeded step

* Fix bundle check

* tryagain

* move readme

* CocoaPods 1.5.0

* Fix generate_template.sh

* Regenerate and test template with generate_template.sh

* Suppress sed error

* Test Xcode 9.2 and 9.3

* Fix 9.3 test

* Clarify README

* Check Xcode version

* Remove Pods folder from template

* Remove Podfile.lock

* Revert "Remove Podfile.lock"

This reverts commit b91fa4083f838878af0fbd57d8ca06dab003cd05.

* Add Swift type checker timer warning

* Add test scheme

* Add shared test scheme

* All the Pods

* Add some options to script

* Remove Circle 1.0

* test inner circleci template

* fix indentation

* require 9.2

* disable 9.3

* verify cookiecutter template has no changes

* fix workspace root

* remove persist_to_workspace

* Fix OOM

* copypaste

* inject SSH key

* fix ssh key

* deploy works

* better destination

* try a two step

* fix commit message

* Move Pods

* Fix Pods out of memory issue

* Fix commit msg

* try something else

+++ generate_template.sh
@@ -1,21 +1,64 @@
+#!/usr/bin/env bash
+
 # This is a shell script to transform the PRODUCTNAME directory into a cookie-cutter template
+
+set -e
+set -o pipefail
+
+# Run this script in its own directory
+SCRIPT_DIR="$(dirname $0)"
+cd $SCRIPT_DIR
+
+echo "Regenerating cookiecutter template from PRODUCTNAME directory contents..."
 
 #This is the only lookup that is done on filenames
 LOOKUP="PRODUCTNAME"
 EXPANDED="{{ cookiecutter.project_name | replace(' ', '') }}"
 
+if [ ! -z "$OUTPUT_DIR" ] ; then
+    echo "Using output directory: $OUTPUT_DIR"
+    mkdir $OUTPUT_DIR
+    cp -rf "$LOOKUP" "$OUTPUT_DIR/$LOOKUP"
+    cp cookiecutter.json "$OUTPUT_DIR/"
+    if [ "${SKIP_REGENERATION}" == "true" ] ; then
+        cp -rf "$EXPANDED" "$OUTPUT_DIR/$EXPANDED" 
+    fi
+    cd $OUTPUT_DIR
+fi
+
+# Clear out any left over artifacts from last regeneration
+if [ "${SKIP_REGENERATION}" != "true" ] ; then
+    echo "Deleting old template output..."
+    rm -rf "${EXPANDED}/"
+    echo "Regenerating template..."
+else
+    echo "Performing dry run on existing template output..."
+fi
+
 # Make the tree
 find ./PRODUCTNAME -type d | while read FILE
 do
     NEWFILE=`echo $FILE | sed -e "s/${LOOKUP}/${EXPANDED}/g"`
-    echo "mkdir -p \"$NEWFILE\""
+    MKDIR_CMD="mkdir -p \"$NEWFILE\""
+    if [ "${VERBOSE}" == "true" ] ; then
+        echo "${MKDIR_CMD}"
+    fi
+    if [ "${SKIP_REGENERATION}" != "true" ] ; then
+        eval $MKDIR_CMD
+    fi
 done
 
 # Copy the files over
 find ./PRODUCTNAME -type f | while read FILE
 do
     NEWFILE=`echo $FILE | sed -e "s/${LOOKUP}/${EXPANDED}/g"`
-    echo "cp \"$FILE\" \"$NEWFILE\""
+    COPY_CMD="cp \"$FILE\" \"$NEWFILE\""
+    if [ "${VERBOSE}" == "true" ] ; then
+        echo "${COPY_CMD}"
+    fi
+    if [ "${SKIP_REGENERATION}" != "true" ] ; then
+        eval $COPY_CMD
+    fi
 done
 
 # Do replacements
@@ -23,9 +66,15 @@ function replace {
     grep -rl $1 ./PRODUCTNAME | while read FILE
     do 
     NEWFILE=`echo $FILE | sed -e "s/${LOOKUP}/${EXPANDED}/g"`
+        SED_CMD="LC_ALL=C sed -e \"s/$1/$2/g\" \"$NEWFILE\" > t1 && mv t1 \"$NEWFILE\""
         # Copy over incase the sed fails due to encoding
         #echo "echo \"$FILE\""
-        echo "sed -e \"s/$1/$2/g\" \"$NEWFILE\" > t1 && mv t1 \"$NEWFILE\""
+        if [ "${VERBOSE}" == "true" ] ; then
+            echo "${SED_CMD}"
+        fi
+        if [ "${SKIP_REGENERATION}" != "true" ] ; then
+            eval $SED_CMD
+        fi        
     done
 }
 
@@ -35,7 +84,38 @@ replace "LEADDEVELOPER" "{{ cookiecutter.lead_dev }}"
 replace "LEADEMAIL" "{{ cookiecutter.lead_email }}"
 replace "com.raizlabs.PRODUCTNAME" "{{ cookiecutter.bundle_identifier }}"
 
-# Delete files that we don't want to include in the template
-rm -rf "${EXPANDED}/app/Podfile.lock"
-rm -rf "${EXPANDED}/app/Pods"
-rm -rf "${EXPANDED}/app/${EXPANDED}.xcworkspace"
+if [ "${SKIP_REGENERATION}" == "true" ] ; then
+    echo "Dry run complete."
+else
+    # Delete files that we don't want to include in the template
+    rm -rf "${EXPANDED}/app/Podfile.lock"
+    rm -rf "${EXPANDED}/app/Pods"
+    rm -rf "${EXPANDED}/app/${EXPANDED}.xcworkspace"
+    echo "Template generation complete."
+fi
+
+# Run Tests
+
+if [ "${SKIP_TESTS}" == "true" ] ; then
+    echo "Skipping tests..."
+    exit 0
+fi
+
+echo "Running tests..."
+
+# cookiecutter default output dir is ProjectName
+TEST_OUTPUT_DIR="ProjectName"
+cookiecutter --no-input --overwrite-if-exists ./
+pushd "$TEST_OUTPUT_DIR/app"
+    bundle install
+    bundle exec pod install
+    bundle exec fastlane test
+popd
+
+if [ "${KEEP_COOKIECUTTER_OUTPUT}" == "true" ] ; then
+    echo "cookiecutter output kept in ${TEST_OUTPUT_DIR}"
+else
+    rm -rf "${TEST_OUTPUT_DIR}"
+fi
+
+echo "Tests complete."
